Reworked flag/flagless comparsion structure.
