+++ include/nil/blueprint/components/algebra/fields/plonk/non_native/detail/comparsion_mode.hpp
@@ -1,0 +1,41 @@
+//---------------------------------------------------------------------------//
+// Copyright (c) 2023 Dmitrii Tabalin <d.tabalin@nil.foundation>
+//
+// MIT License
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//---------------------------------------------------------------------------//
+
+namespace nil {
+    namespace blueprint {
+        namespace components {
+            namespace detail {
+                // Ordering is important here: the logic differs for FLAG and non-FLAG modes.
+                // We use comparsion with comparsion_mode::FLAG in order to distinguish between the two cases.
+                enum comparsion_mode {
+                    FLAG,
+                    LESS_THAN,
+                    LESS_EQUAL,
+                    GREATER_THAN,
+                    GREATER_EQUAL,
+                };
+            }   // namespace detail
+        }       // namespace components
+    }           // namespace blueprint
+}   // namespace nil
